适配新livox驱动; add localization_status ; disable c++17 std; rm pangolin, opencv; rm control_status_

+++ src/include/livox_datatype/livox_ros_datatype_def.h
@@ -34,7 +34,8 @@ using ImuMsg = sensor_msgs::Imu;
 
 using PointCloud = pcl::PointCloud<pcl::PointXYZI>*/
 struct LidarPoint{
-    u_int32_t offset_time;//      # offset time relative to the base time
+    // u_int32_t offset_time;//      # offset time relative to the base time
+    double offset_time;//      # offset time relative to the base time
     float x;//               # X axis, unit:m
     float y;//               # Y axis, unit:m
     float z;//               # Z axis, unit:m
